📝 (adr) Add ADR 0003: content organization and accesses
This ADR describes how videos are organized in playlists and how these
playlists are managed.

There is a special section "ongoing discussion", present only for the
review/discussion process. Once the decisions will be taken, this
section will be removed and the final choices will be integrated in the
"Decision" section