✨ Enhance "validate_unique" to check "NonDeletedUniqueIndex"
As we don't use `unique_together` because of the soft-deletion, the
default `validate_unique` method of models does nothing.

So we enhance it to simulate `unique_together` by extracting fields from
our uses of `NonDeletedUniqueIndex`.

This is a lot helpful in the admin because we can have a meaningful
error, by Django, instead of integrity errors by the database.