♻️(models) refactor models/factories/admin
I think we should not continue using types on this project
because it adds a lot of complexity and a barrier to entry
with limited benefits. I think types on a Django project
will be interesting and efficient when the Django codebase
will integrate types. In the same operation, I refactored
models to be a bit more generic and extensible.

The main change is to adopt Github's pattern for user accesses
ie use a unique model OrganizationAccess (resp ConsumerSiteAccess
or PlaylistAccess) to define a user's access to an organization
instead of a seperate model for each type of access.