♻️(playlist) change relation between video and playlist
Allowing a video to belong to several playlists complicates
everything. For LTI, it is better if videos directly
identify as unique resources. If a video needs to be
associated with another playlist, we can copy it instead of
adding many to many relations. This will also allow to
customize a video in a playlist independently of the
original video.