✨(front) define types for our application data
Our application data (what amounts to a global state in our small
app) is structured and pretty specific. We can define what it looks
like so as to trivially propagate it as we see fit.

Note: we don't perform checks on it for now but we could easily add
type guards to ensure correctness at bootstrap time.