♻️(models) rename "name" field to "title" for video and playlist
The LTI standard uses the "title" fields to refer to human readable
names of a resource_link and context. It is thus clearer if our video
and playlist have a "title" field instead of a "name" field. the "name"
field is kept for machine name of on the consumer site.