✨(API) add an API endpoint for the video object
Authentication is done via a JWT token linked to the video so that
we don't need to create user accounts. Instructors are able to
update the name and description of their video via this endpoint.
Amazon AWS will also update the state of the video after it has
finished encoding it.