✨(front) complete AppData type with a Video record
The Video record information is available at load time as it comes
from the HTML in the initial template provided from the server. We
can therefore depend on it being present at all times (after load)
and containing the necessary data.

On the other hand, upload policies are not part of the initial payload
and need to be separately fetched later on. They thus cannot be part
of AppData (and should be removed from our template data elements).