✨(front) add an ErrorComponent to handle error cases
We need to be able to display relevant error messages in case
we encounter LTI errors or a student tries to access a video
that is not published yet and receives a 404.

Over time we can add more cases for this component. We'll also
style it to fit and look clean in the iframe space.