🔧(tslint) update tslint rules for prettier & style consistency
tslint recently changed the default (or added a default) for property
names in objects. This clashes with the standard way to write JS and
with what prettier is enforcing. Change it to only allow quotes when
they are strictly necessary.

Also remove the one-line rule for if/else clauses, which makes no
sense when writing branching logic with comments that can't be
coerced easily into a switch.