🐛(video) allow POST calls on the /lti-video view
The /lti-video view is _expected_ to be called by POST requests
as it needs arguments contained in the POST body to operate.

TemplateViews on the other hand, while they are a useful shortcut,
can only support GET requests.

Refactor the /lti-video view to use a generic View with just the
TemplateResponseMixin so we can enable POST requests and still simply
produce a templated response.

NB: disable CSRF protection as this view is only supposed to be called
directly through iframes with no prior request to our Django server, and
uses other security protocols specified by LTI.