✨(lti-video) use data-elements to populate the App data context
In order to structure our app properly, we'll have to use routing
mechanisms and organize our component structure. This makes it
impractical/impossible to just pass around the whole app state to
be consumer wherever parts are needed.

Instead, we can rely on an AppData context to make appData available
to components, however deep in the tree, that need it, without polluting
intermediate levels.

To export it, we must place it out of the DOMContentLoaded callback. We
therefore need to include the script at the bottom of the <body>, rather
than in the <head>.