🔧(front) add a simple mock for react-intl in our unit tests
Wrapping all components we test that contain strings with an
IntlProvider and having to account for it in our tests was getting
old really quickly.

Write a simple manual mock of react-intl to allow us to stop using
it in our tests. It still lets us test our strings as it hands back
and prints default messages.