✨(front) implement the video uploading form
VideoForm is the core of marsha's frontend functionality: it allows
privileged users (eg. instructors) to upload a new video and add
a title & description from the lti-video view, which is supposed to
be called in an iframe.

Get the policy from the server in order to upload the file
directly to AWS s3.

NB: title and description are not used for now.