✨(front) add a flushAllPromises helper for our unit tests
Whenever we use promises in our code, we're bound to require a
helper to flush outstanding promises to enable us to run unit tests
and check values easily.

We decided to put it in testSetup as it is a file we import at the
top of each of our tests (for its side effect of enabling enzyme) and
we will not need an additional import statement to take advantage of
this helper.