✨(terraform) add a mediaconvert preset for frame capture
Capturing a thumbnail to display as a placeholder for the video
and displaying small preview captures at each second of the
video are two different things. For thumbnails we only sample
one image in each resolution. For previews, we sample one image
per second on the whole video but in a size of  100 pixels.