✨(terraform) add CloudFront behavior to restrict access to mp4/subtitles
We want to restrict access to videos and subtitles but it is harder to protect
DASH, HLS and preview images because they have numerous files which prevents us
from using signed urls. We would need to use signed cookies but it requires a
custom Cloudfront domain to avoid doing CORS in iframes. In the meantime we
propose to only restrict access to mp4 videos and subtitle tracks.