✨(front) replace empty RootComponent with our Router
As long as we don't have too many routes, it's simplest to just keep
them listed in AppRoutes as a central directory that encompasses the
whole app.
If it gets out of hand, we'll look into ways to move the routes to
Routable component definitions and dynamically load them in the
router.