✨(backend) allow activating signed urls on the API
The video API returns the urls of the video and thumbnails in
AWS (hosted in S3 and accessible via CloudFront). This commit
allows to return signed urls so that we can make our videos
private in CloudFront.

Note: For the moment, we do not expose preview imagess (they are
already computed in AWS so that it will be easy to add them later).
It will require using signed cookies as there are too many
preview images to be able to compute a signed url for each of them.
We think it is not straight forward to make videojs add a cookie to
its requests to CloudFront's domain and make it work for all browsers.
A better way to do it is to use a custom domain in Cloudfront so that
there is no CORS issue and the cookie is sent automatically by the
browser. We think that setting this up would delay the project and
prefer to leave it as an future improvement.