✨(front) use frontend URL fragments from variables
Using strings directly wherever we define a route or link/redirect to
them can cause issues when some of the URL fragments change or when
we decide to translate them.
Using functions returning strings instead gives us a simple API that
manages these issues and will enable us to add a helper to enable
URL localization later on if we need to.