✨(front) pass the JWT in AWS S3 file metadata
We need to update the video status from a lambda once it
has finished processing in our mediaconvert pipelines. To
enable the lambda to do this update, we need to pass it
the JWT that allows the video record to be updated in our
django app.

The simplest way to do that is to just add it as metadata that
will automatically be available in the relevant lambda(s).