✨(video) allow updating a video by uploading a new file
We want an instructor to be able to update an existing video
by uploading a new one just the way he uploaded the first one.

In the previous implementation, the destination file names
only referred to the video and playlist ID so the files would
be overwritten upon transcoding and the cache in CloudFront
would hide the change.

We now generate a new stamp each time the client requests an
upload policy and include it in the key of the files in S3. This
stamp will also appear in the destination files created by AWS
so that the cache in CloudFront is invalidated. If the upload
succeeds, AWS will update the "uploaded_on" field on the video
so that the new urls are now returned by the API.