♻️(front) use MemoryRouter instead of HashRouter for frontend routing
We used HashRouter to manage our frontend routing. This allowed us to have
a proper history (even though hidden away by the iframe) as we move around
the app.

It is however not ideal with LTI. In some situations we need to reload the
iframe, which does not work if we navigater in the browser's history as it
then sends a GET with the new URL instead of POSTing the original one with its
params.

Using MemoryRouter solves the issue as it does not touch the browser's
history at all. We therefore remain on the same URL and can reload our
POST request with its params.