💄(front) add a styled compatible CSS reboot
We need to use a CSS reboot to start from a clean slate and minimize unwanted
differences between browsers.

We picked Bootstrap's as it's pretty vanilla and achieves the look and feel we
want. styled-reboot simply ports it to styled-components so we don't have to try
and load it separately.