✨(front) add a typing for styled component wrapper
styled-components supports adding properties to basic components
on-the-fly. Its typings however do not describe this feature.

It's useful for eg. making a Button component that is a styled button,
that accepts a "variant" prop so we can have different colors from a single
component.

This typing can be used to enable that on specific instances, by simply passing
an interface listing the available props & their values' types.