✨(front) improve the UI for the VideoUploadField
We want our VideoUpload field to emphasize that it works
as a feature and users can just drag-and-drop files onto it,
and still keep a real button for the file input for accessibility.

VideoUploadField is mostly a functional container, DropzonePlaceholder
is the visual aspect of the dropzone including the button, and
DropzoneIcon is an SVG icon for the dropzone.

We added reusable styles for a Button component as we needed a styled
button for the DropzonePlaceholder.