♻️(front) add a loading state to VideoForm
When the user selects a file in VideoUploadField, we immediately
trigger the upload. Some of those video files can be pretty large,
which means we need a loading screen for the duration of the uploads
so users know what is going on and don't eg. accidentally cancel the
upload.

Use a simple CSS animation for now to show this loading state.