✨(terraform) add necessary outputs to Terraform
We need to output the data that Marsha will need:
access key id and secret key for the AWS account
used by Marsha to upload videos. The Terraform
module that creates the state bucket also needs to
output the KMS key id that the main Terraform module
uses to encrypt its state.