♻️(database) make database settings more flexible
In Arnold, the different database parameters are handled separately.
For example, the password is stored in a secret and mounted as an
environment variable in the pod where as the host is not secret and
changes upon each deployment. We applied the same pattern as what is
done in github.com/openfun/richie