✨(front) extend react typings for the <source> element
We are using <source> elements inside <video> to list all
available sources that the player can load depending on the
device/network/user choice.

To display a source selector, Plyr (our new player of choice)
uses a "size" attribute on <source>, which is not a native HTML
attribute. Extend react's definition for the default JSX elements
so it is accepter by TypeScript as a legitimate attribute.