✨(subtitles) add version fields to subtitles to allow updating them
Similar to what we have done on the video object, we want to be
able to update subtitles by uploading a new file to S3 so we
need to add a stamp to the name of the destination file in
order to invalidate CloudFront's cache. The "uploaded_on" field
records the time at which the upload policy was retrieved and the
"state" field is where AWS reports the status of the latest
upload by patching both fields simultaneously.