✨(subtitles) add API endpoint for subtitle tracks
This endpoint works a lot like the video endpoint. For authentication,
It expects a JWT token related to the video linked to the subtitle
accessed. As opposed to videos, subtitle tracks can also be created
via the API as long as we have a valid JWT Token for the related video.