🐛(videos) fix broken tests and add missing tests for the video API
Some old code was forgotten tests (we did not notice because they were
failing for the wrong reason). Some tests were also missing to validate
that the fields can be patched on the video API endpoint. We need the
patch method to make sure that AWS can update the state of the object
without interfering with eventual updates that the user could make
simultaneouly to the name or description of the video.