✨(terraform) add a policy to allow lambdas to read source s3 bucket
Our lambdas (only `encode` at first but maybe others later on) need
to be able to read metadata on existing objects in the relevant S3
source bucket.

There is no specific permission for metadata, so we need to give them
read access to those objects. The way to do that is to target the
marsha_source by its arn and then select all objects inside it.