🚸(tooling) add a development view to simplify LTI based work
LTI providers need to respond to a POST request on a route
with a specific set of parameters. We also intend it to be used
through an iframe.

We therefore need a development view to be able to make multiple
POST requests hassle-free during development. Developing in an
iframe is also more comfortable as it mimicks the environment
the real app will run in.