✨(upload) add an endpoint for AWS to confirm uploads
This endpoint expects a object "key" and parses its elements to
retrieve the object type, object id and stamp. It then updates
the active stamp on the targeted object in database and its state
so that it becomes available (if ready) or is marked as a failed
upload so that the client knows he needs to try again.