🔥(policy) do not pass client jwt as video metadata
We planned on sending the client JWT as part of video metadata to
enable our lambdas to use it to update the video on the API when
upload & transcoding are finished.

This design was scrapped in favor of one where we actually use a
shared secret to authenticate the lambdas with the backend server.
We therefore should not send the client JWT to S3 anymore.