✨(terraform) add environment variables for the confirm lambda
In order to implement the confirm lambda (that calls a Django
route when processing of a video is complete), we need it to
have access to the Django endpoint URL, and a shared secret to
sign messages.

We can just pass the endpoint URL as an environment variable. As
for the shared secret, we're generating it from a deployment
stamp (using terraform's random provider) and adding it to our
outputs so it is not lost to us. It's the deployer's responsibility
to pass it to Django as an environment variable.