⏪(terraform) revert "add a policy to allow lambdas to read sou..."
This reverts commit 11a46d4e5d21540c2b9b3ca9302073eed73ce7ed.

We decided not to use metadata to transmit a JWT to the lambda and
instead share a proper secret between the lambda and the Django
backend.

It is therefore not necessary for lambdas to have read access on
our source s3 buckets; we should not grant privileges that are not
actually required. We can just revert that permission.