✨(terraform) add relevant environment variables for confirm lambda
The confirm lambda needs the URL for the endpoint it needs to call
to confirm upload & transcoding status. It also needs access to a
shared secret that will authenticate its requests by providing a
signature.

We're simply passing both of those as environment variables into
terraform (and the shared secret as an environment variable to Django
as well). This allows us to change the secret without downtime.

We can have a Django app up-and-running with the new secret when we
update the AWS infrastructure. This would not be possible if we let
terraform generate the key itself.