✨(api) authenticate upload confirm request using shared secret
Use shared secrets provided through the environment to calculate
hmacs of the s3 object key. This allows us to ensure unauthorized
parties cannot make accepted request to update a file status on
this API.

We use a list of secrets so we can deploy a version of the API that
supports the existing environment AND an upcoming environment that
will use another key. We can then painlessly switch AWS environments
knowing our backend is ready for the change.