✨(lambda) implement the upload confirm call triggered from AWS
When a file has finished upload and processing, the confirm lambda
is triggered by our terraform infrastructure. Its job is to sign
the S3 object key with the secret is shares with our Django backend
and post the key + signature + state payload to it.

Also add some dependencies to enable testing using Jest.