👷(circleci) run tests for the confirm lambda
Add a job and update the marsha workflow in the Circle CI config
file to run our new confirm lambda tests. Update the package.json
script to run using jest.

Note: we chose to repeat dependencies and CI config in order to
ship now. We should consider a refactor to streamline, complete
and generally improve our JS testing to bring lambda testing &
linting up to par with what we're doing for the frontend.