🐛(lti) enforce consistent type for the video "urls" fields
The Video "urls" field as served by the LTI view is expected to
be JSON. This enables the client to simply use JSON.parse and
use the results.

Here the return value was JSON *when there was a value* and just
`"None"` as a string otherwise. Use json.dumps to serialize even the
None value to the JSON `"null"`.