✨(lti) allow portability of videos between playlists and consumer sites
A course may be copied in the LMS which results in the LTI launch request
targetting a video with a known "lti_id" (that we already have in database)
but with a different playlist and/or for a different consumer site. In this
case, we must be able to automatically duplicate the existing videos and
reuse the same video files. By default, we allow videos to be portable to
another playlist but not to another consumer site.