♻️(front) rename Video property from "status" to "state"
We implemented our Video type in the frontend before this property
was available on the backend and we provisionally named it "status".
It's actually named "state" on the backend.

Just rename it on the type and at all use sites to conform to the
actual shape of a Video object.