✨(terraform) update transcoding presets/jobs for adaptive bitrate
We previously went for separate DASH & HLS presets & jobs that
produces different sets of files. This is not a necessity.

Using CMAF, we can run one job with its presets, and produce a
set of files that is compatible with both DASH & HLS. This way we
cut our encoding & storage use (and costs) and need less code to do
it! Amazon generates both manifests itself thanks to flags in the
Job produced by the encode lambda.

Note: multiplexed files are not supported by adaptive bitrate formats,
which means we had to separate audio & video outputs and therefore
write & register separate presets for them.