➕(front) add shaka-player along with minimal typings
We need a JS player to be able to play videos using DASH ISO
for adaptive bitrate in browsers.

We chose shaka-player as we encountered some issues with the
Dash Industry Forum library, notably dashjs does not adapt between
audio adaptation sets of the same language (as generated by AWS
MediaConvert) whereas shaka-player does.

Unfortunately, shaka-player does not provide TS typings, which
is why we started up our own typings file, to be completed later
on and folded into DefinitelyTyped.