♻️(api) rename upload-confirm to update-state to handle PROCESSING state
Add a new "processing" state for objects that have finished uploading but are
undergoing additional automated processing.

The upload-confirm API was already not just a confirming API as it could
be used to update a status to "error". Now that we're adding another
additional state, it makes sense to rename the endpoint to better
represent the breadth of the operations it can perform.