♻️(lambda) make updateState lambda code easier to reuse
We're currently lacking an acceptable way to reuse code accross
lambdas (apart from having them call each other, which incurs
infrastructure costs)

As a result, we've decide to just duplicate the code we need
until we have some time to implement a better solution. Here,
we're separating the part of the code we want to duplicate so
it's easier to copy it and to refactor once we figure out how to
reuse code.