✨(lambda) add an update-state call for video processing
Some videos can take a long time to process (transcoding an hour
long video can take one hour). This means we can't expect users to
stay on the same page during that time: we therefore need a transition
state between "pending" (there is no video) and "ready" (there is a
video).

This way, when users come back to the page or reload the iframe, we
can inform them of the actual state of the upload and show them a
useful message so they don't attempt to re-upload the video.