♻️(lambdas) update terraform Makefile to stop using NPM
The lambda-related tasks in the terraform Makefile were using
NPM instead of Yarn to install dependencies & run tests. This
resulted in generated package-lock.json files in the repositories,
which are irrelevant as we're using Yarn and cause warnings to be
emitted.

Just use Yarn instead. Remove existing node_modules and use
--frozen-lockfile & --production to make sure the generated bundles
include the correct dependencies and no dev dependencies.