♻️(lti) stop considering hostname from the LTI query
It seems that the Open edX LTI Xblock does not always send the
hostname, for example when calling the LTI Xblock from the CMS.

This causes Marsha to refuse the LTI connection because the passport
can not be found. Since the consumer site is already defined by the
passport, we can skip checking the hostname and rely on the consumer
site linked to the passport after we have retrieved and verified it.
This method will also be used for other LMS like Moodle in order to
stay coherent.