➕(front) add styled-components helpers for Jest
"jest-styled-components" includes helpers for Jest testing. Some of
them are about snapshots (which we're not using right now) but the
most interesting one for us is "toHaveStyleRule".

This lets us check variable CSS props which change depending on
input/context. We can also use that to ensure some of the most
important CSS rules are applied later on.