✨(front) expand video state list to include UPLOADING & PROCESSING
PROCESSING is a new state that exists on the server side along with
PENDING, READY and ERROR. We need to add it here to have the complete
list.

As for UPLOADING, it does not exist on the backend side, as during
upload the model value for the state is still PENDING. However, for
us on the frontend, they are two different states of the app & the
interface.

We can therefore add it to simplify our code and make it easier to
work with those states. This state will not actually be applied on
any Video but will be used as a stand-in wherever we're passing
video state without the full video and need to represent this state.