✨(front) add a set of components to display an upload's status
In our dashboard (or elsewhere, later on), we'll need to display the
status of an upload along our pipeline, from the initial PENDING
to READY or error through all our steps.

This component set provides a simple element to show the current
state and where it's located in the pipeline through a single
property (the state as videoState).