✨(front) add a higher-order component to route to a view
react-router does not provide any straightforward way to imperatively
navigate to a route as needed.

This can be needed in some circumstances, most notably when we want
to use something else than an <a> element as a link, and style it
at the same time. react-router & styled-components do not play nice
together as we would end up with either dropped styles, an extra
element in the DOM, or broken types.

With this helper, we can easily add a navigating onClick handler to
any element, including a styled element.

NB: types may be imperfect as this was somewhat clunky to type, and
tests are missing as they were too challenging to write in a reasonable
amount of time.