✨(front) add an App component to allow video updates
We initially put appData straight into a Context, which allowed
us to pass it through easily to any child that needed it.

However, it came from a static value in index and could not be
modified. We need to put it in a component's state to be able to
change it and have the changes be reflected in our renders.

This is the role App fills. If/when we need to support more changes
than just updating the video, we'll have to consider bringing on a
state management solution or building our own. For now, we think it
would add unnecessary complexity to a relatively simple app.