✨(front) mark the video as PROCESSING when upload is finished
When a video is done uploading, we need to immediately mark it
as PROCESSING locally.

That is because that is the next logical step, and the server will
not be aware of this change until the "encode" lambda updates it
through an HTTP request.

Rather than wait on this to happen and poll the server, we can
optimistically update the state and leave it to the Dashboard to
do some polling to keep it up to date later on.