✨(front) add a Dashboard component to show video state
We need a view to show video state through upload & processing. We
previously used a simple loader for the uploading state but that is
proving insufficient as processing can take a long time and we need
to inform the user.

As we're going to need a dashboard anyway (to monitor/replace uploads
and manage tracks such as subtitles), we can start with a simple
version of the Dashboard to show video states and update it later on
when we have more information to show there.