🔧(front) add a basic mock for react-router in our unit tests
withRouter (and therefore our own withLink) need to be run inside
components that have a Router in their ancestors. This is not a
problem in app code as we have a Router near the Route. In tests
however, it would force us to wrap some of our tested components in a
Router.

Wrapping all components that use withRouter or our own withLink in
a Router is problematic because:
- it's clunky;
- it prevents us from using shallow (as only the Router is rendered then
  and not the component we're trying to test!).

This is why we decided to start our own auto-mock for react-router. For
now we just have to do an implementation of withRouter as we do not render
any other components from react-router in our unit tests.