✨(front) update AppRoutes & RedirectOnLoad to use Dashboard
We need to add our Dashboard component to AppRoutes to be able to
redirect or use links to it.

There's also a case where the correct redirect at startup is to the
Dashboard: that is when the instructor is loading the view and there
is an existing video (not PENDING) that is also not READY.

Update tests to account for this case and use the proper Enum for
video state names.