✨(front) have VideoForm direct users to Dashboard
As we added a more complete Dashboard, the existing UploadingLoader
is not useful anymore.

Instead, we can use a small trick: while we're redirecting to the
Dashboard after the upload is complete, we can't do that for the
UPLOADING state as it would unmount our form component.

Instead, we can just have VideoForm render the Dashboard in the
UPLOADING state and just redirect when the upload is done and the
state moves to PROCESSING. Apart from the state change (which is
intentional), the redirect is invisible for the user.