🚚(project) move frontend-related files to `src/front`
marsha's file structure is currently a little bit complicated, with
a lot of directories and files straight at the root. This can make it
complicated to approach for new developers.

To begin to remedy this, move all frontend related files to src/frontend,
which should now look like a regular frontend project directory.