🚚(project) move terraform & lambdas to `src/aws`
As we continue reorganize the project's files, we start by moving
terraform into the `src` directory. This directory was already
self-contained, which means we did not have any changes to make.

We just renamed it so `aws` so it's easier to understand what it
does at a glance, and put the lambdas directly in it for clarity.

This led us to put `.prettierrc` in `src` too so it is shared between
all our JS-using projects.