🚚(project) move terraform & lambdas directly into src/
As we continue reorganize the project's files, we start by moving
terraform into the `src` directory. This directory was already
self-contained, which means we did not have any changes to make.

We just moved the lambdas to their own folders in `src` as each of
them is a service & a nodeJS project of its own.

This led us to put `.prettierrc` in `src` too so it is shared between
all our JS-using projects.