👷(project) enforce presence of all authors in CONTRIBUTORS.md
In Github issue #91, @xanderyzwich suggested we should add a tool to
our build to make sure our list of contributors in CONTRIBUTORS.md
is always up-to-date. This is a tentative to add such a check via
a command line run in CircleCI that fails if the author of one of
the commit contained in a PR is not included in the CONTRIBUTORS.md
file.

Fixes #91