🐛(devtools) add an oauth_consumer_key field to the development view form
After changes in the way informations from the LTI request are collected
and used, the Django backend is now using the consumer key to
retrieve the passport, and, from there, the consumer site.

This means we cannot just bypass LTI checks in development and hand
over a consumer site that will work out of the box. We need the LTI
request to pass an oauth consumer key that matches to a real passport
with a linked consumer site in the database.