🐛(lti) fix marsha when placed behing a TLS termination proxy
The LTI signature is computed using the url of the LTI launch request.
When Marsha is placed behind a TLS termination proxy, the url as seen by
Django is changed and starts with the "http" scheme instead of "https" in
the consumer. We need to revert this so that the signature we calculate
matches the one calculated by our LTI consumer.

Note that this is normally done in pylti's "verify_request_common" method
but it does not support WSGI normalized headers so let's do it ourselves.