📝(env) document available environment variables for Django & TF
We're in the process of making the project more accessible to
newcomers. One of the pain points we identified was that there
are a lot of required environment variables in both Django &
Terraform, and that they are not documented anywhere.

In addition to the development env template we recently added,
we're thoroughly documenting all our environment variables.