♻️(settings) don't sign Cloudfront URLs by default in deelopment
We recently changed `CLOUDFRONT_SIGNED_URLS_ACTIVE` to be able to
set it through an environment variable. This makes marsha's settings
more consistent as most of the customizable behavior can be configured
through the environment.`

In our opinion, `CLOUDFRONT_SIGNED_URLS_ACTIVE` should be False by
default in Development to make onboarding easier for new users
and developers on marsha.