🐛(docker) fix application workdir
Container CMD (running gunicorn) was broken as the wsgi script moved to
src/backend/marsha. We've restored the workdir to the Django project
root as expected for a Django application.

Calls to management commands have been fixed consequently.

We also fixed the front-end build path that would not have been
collected (as it was not in the Django app statics).