🐛(front) fix an issue with VideoPlayer when the video is first ready
When the upload is ready at the end of processing, the instructor can
click on a "Watch Video" button to go to the player. This is broken
right now as the video object `<VideoPlayer />` receives from the App
state is not up-to-date.

Add a call to App's `updateVideo` in `<Dashboard />` polling to
ensure App state contains the most up-to-date video, including the
URLs.