✨(front) enable automatic iframe resizing through iframe-resizer
In order to be able to display marsha's content in the best possible
way, we need the host page to dynamically resize the iframe based
on its content.

This enables us to add UI elements (like transcripts, download links,
instructor-specific UI) without having to shrink the video or form
too much.