✨(front) rework frontend layout to work with iframe-resizer
iframe-resizer works by measuring the contents of the iframe and
adapting the iframe to its measurements. Obviously, this cannot
bring us any results when our layours are built to occupy 100%
of the iframe height all the time.

We can instead base them on an aspect ratio. As we're working with
video, we can take 16/9 as a starting point, which we can easily
achieve through the use of the `vw` unit for both width and height.

This gives our layouts are fixed height, which iframe-resizer can
then use as a target to adapt the iframe itself.