🔊(lambda) add logs before update-state requests are fired
The `encode` and `update-state` lambdas fire requests to the Django
backend, using a shared secret to authenticate the requests. Those
requests are *not* supposed to fail unless the Djago server is down
or having issues.

Add logs of the outgoing requests' method, URL and body to help
with debugging when they do fail.