🐛(lti) fix checking what the LTI "verify" method returns
The "verify" method on the LTI class used to return True or raise
an exception but it now return a the consumer site object when it
passes. Tests had not been updated because they were still passing
but we can improve them by actually checking that the consumer
site object is returned as expected.