✨(front) implement the instructor view wrapping <VideoPlayer />
As a 1st priority, we need to give instructors a way to change the
video they uploaded for any given resource, that is the video that
will be displayed in the corresponding location in the LMS.

To achieve this, we decided to implement a more generic <InstructorView />
that we'll be able to use for other features later on. We also used
an <InstructorWrapper /> that lets us automatically add the
<InstructorView /> on any view where it is needed, only when the current
user is an instructor.