♻️(subtitle) refactor to add transcripts
Transcripts are subtitle files (srt, vtt,...) that we display outside
of the player, below the video and that scroll in synchronization with
the video. Since they are technically the same thing as subtitle, we
propose to use the same object.

Another type of subtitle is closed captioning and there was already a
flag on the SubtitleTrack object for that. Since closed captioning and
transcripts are mutually exclusive, we decided to refactor this flag
and make it a char field called mode that can be equal to:
- None: normal subtitle included in the video,
- "ts": transcript displayed outside of the player,
- "cc": closed captioning included in the video.

A video can simultaneously have all 3 modes of subtitles for a given
language.