🐛(api) fix the update-state request serializer
The update state request serializer uses a regex to validate s3
object keys. It had not been updated to account for changes in
the way timedtext files are stored in s3.

Updating the regexp to make the "mode" part of the object key
mandatory and accepting all valid modes (`"st"`, `"ts"`, `"cc"`)
fixes the issue.