✨(lambda) update the encode lambda to process timed text
The encode lambda was originally written to handle incoming
"create object" events for videos and trigger a MediaConvert
job to encode them in all the formats we need.

As we add support for various forms of timed text (subtitles, ccs,
transcripts), we need to handle those as well. This can be done
directly in JS in the lambda as it is much less demanding than
converting video.

We reworked the lambda to convert all incoming subtitles to WebVTT
and write them to the destination bucket. To do this, it needed
additional access rights to s3.