♻️(account) add a domain field to the ConsumerSite model
The ConsumerSite model only had a "name" field in which we were
recording a domain name. This was confusing. Following what is
done is Django for the site contrib, we added a `domain` field
separate from the name field.

Note: we had to add the site contrib to installed apps because
Django was asking when we try to import the domain validator from
the contrib site models file.