🎨(video) remove the language field on the Video model
A video can multiplex several languages so this field should be
called languages and expect a list of languages multiplexed in
the video. We are not supporting this for the moment and only
kept the first languages while encoding. Let's just remove the
field for the moment and we will add a `languages` field again
as soon as we improve our encoding to retrieve the list of
languages from the video.