🐛(videos) fix a 500 error when serializing videos
We encountered an issue where attempting to serialize videos with
an associated TimedTextTrack would fail due to the
TimedTextTrackSerializer keeping the UUID as an UUID instance during
serialization instead of converting it to a string.

The string conversion can be forced by explicitly specifying the field
content (using `PrimaryKeyRelatedField`) which lets us control how
it is serialized.